fixes for pep8, local import for GPKernel, and moving ovr_decision_function to utils
