+++ sklearn/model_selection/tests/test_split.py
@@ -25,7 +25,6 @@ from sklearn.utils.testing import assert_raise_message
 from sklearn.utils.testing import ignore_warnings
 from sklearn.utils.validation import _num_samples
 from sklearn.utils.mocking import MockDataFrame
-from sklearn.utils.metaestimators import _safe_split
 
 from sklearn.model_selection import cross_val_score
 from sklearn.model_selection import KFold
